// frontend/src/generated/SubsetSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NodeHandle } from "./NodeHandle";
import type { PredicateDistribution } from "./PredicateDistribution";
import type { SubsetCounts } from "./SubsetCounts";
import type { SubsetFilter } from "./SubsetFilter";

export type SubsetSnapshot = { predicates: Array<SubsetFilter>, visible_nodes: Array<NodeHandle>, visible_edge_ids: Array<number>, counts: SubsetCounts, distributions: Array<PredicateDistribution>, };